+++ .gitignore
@@ -1,5 +1,7 @@
 bot-env/
 __pycache__/
 matches/
+.*/
 
+riotgames.pem
 secrets.json
