+++ scripts/classes/ClientNotOpenException.py
@@ -1,0 +1,2 @@
+class ClientNotOpenException(Exception):
+    pass
