+++ scripts/classes/MatchHistoryMatch.py
@@ -1,0 +1,6 @@
+from classes.PlayerGameStats import PlayerGameStats
+
+class MatchHistoryMatch:
+    def __init__(self, gameStats: PlayerGameStats, mmrDelta: int) -> None:
+        self.gameStats = gameStats
+        self.mmrDelta = mmrDelta
