+++ scripts/classes/Teammate.py
@@ -1,0 +1,18 @@
+class Teammate:
+    def __init__(self, playerDisplayName: str) -> None:
+        self.playerDisplayName = playerDisplayName
+        self.wins = 0
+        self.losses = 0
+        self.gamesPlayed = 0
+        self.winrate = 0
+
+    def add_game(self, win: bool) -> None:
+        if win:
+            self.wins += 1
+        else:
+            self.losses += 1
+        self.gamesPlayed += 1
+        self.winrate = int(round(self.wins / self.gamesPlayed * 100))
+
+    def __repr__(self) -> str:
+        return f"{self.wins}W {self.losses}L ({self.winrate})"
