+++ scripts/utils.py
@@ -1,4 +1,3 @@
-
 def chunks(arr: list, n: int):
     '''
         returns arr broken into n-sized chunks
